Refactor: sample api files
