--- src/main/java/com/yjp/delivery/controller/sample/dto/request/package-info.java
@@ -1,1 +1,0 @@
-package com.yjp.delivery.controller.sample.dto.request;
