--- src/main/java/com/yjp/delivery/controller/sample/dto/response/package-info.java
@@ -1,1 +1,0 @@
-package com.yjp.delivery.controller.sample.dto.response;
